add scetrca v2.0
